requirements.txt
fastapi==0.115.6
uvicorn==0.38.0

openai==1.61.0
numpy==1.26.4
faiss-cpu==1.9.0.post1

python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-dotenv
